main.py
from flask import Flask, request, Response
import telebot
import boto3
import google.generativeai as genai
import logging
from io import BytesIO
import PIL.Image

app = Flask(__name__)

# Set up logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Telegram Bot API token
bot = telebot.TeleBot("TELEGRAM_BOT_TOKEN", parse_mode=None)

# Generative AI API key configuration
genai.configure(api_key="GENAI_API_KEY")

# Set up the model
generation_config = {
    "temperature": 0.9,
    "top_p": 1,
    "top_k": 1,
    "max_output_tokens": 2048,
}

safety_settings = [
    {"category": "HARM_CATEGORY_HARASSMENT", "threshold": "BLOCK_MEDIUM_AND_ABOVE"},
    {"category": "HARM_CATEGORY_HATE_SPEECH", "threshold": "BLOCK_MEDIUM_AND_ABOVE"},
    {"category": "HARM_CATEGORY_SEXUALLY_EXPLICIT", "threshold": "BLOCK_MEDIUM_AND_ABOVE"},
    {"category": "HARM_CATEGORY_DANGEROUS_CONTENT", "threshold": "BLOCK_MEDIUM_AND_ABOVE"},
]

model = genai.GenerativeModel(
    model_name="gemini-1.5-flash",
    generation_config=generation_config,
    safety_settings=safety_settings
)

convo = model.start_chat(history=[])

# AWS Rekognition client
rekognition_client = boto3.client(
    'rekognition',
    aws_access_key_id="AWS_ACCESS_KEY_ID",
    aws_secret_access_key="AWS_SECRET_ACCESS_KEY",
    region_name="AWS_REGION"
)

@app.route('/', methods=['GET', 'POST'])
def index():
    # Implement the chatbot logic here
    pass

if __name__ == '__main__':
    app.run(host="0.0.0.0", port=5000)
